requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg[binary]
redis
celery
firebase-admin
